Added the ability to add courses and exposed via a router
It was tough to wire up a basic router because Koa functions very differently than expresss.
Because it uses generators, all promises need to be __yielded__, otherwise it will simply
ignore them. This isn't obvious, but it makes the code easier to read (yield means that
we are waiting for an asynchronous action, no guessing), and Koa's style for handling
middleware and writing requests seems more sane.